requirements.txt
ultralytics
fastapi
uvicorn
opencv-python-headless
pillow
